go network: add listener
